+++ rust-script/tests/macro_test.rs
@@ -4,8 +4,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-godot_rust_script::setup!(tests_scripts_lib);
-
 #[test]
 fn verify_macros() {
     let _ = || {
